Update Cali geojson quadrants and properties
Reorganized and renumbered quadrant features in cali_base.geojson and cali_base_p.geojson, updating coordinates, codes, and color properties for consistency. Also added record count to CSV download help text in app.py and created a backup for geojson_standardizer.py.
